+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [#119](https://github.com/dweymouth/supersonic/issues/119) Add file path column to tracklist view
 - [#117](https://github.com/dweymouth/supersonic/issues/117) Add (optional) system tray menu and close to tray support
 - [#55](https://github.com/dweymouth/supersonic/issues/55) Show disc number and disc count for multi-disc albums
+- [#115](https://github.com/dweymouth/supersonic/issues/115) Add search bar to artist, genres, and playlists pages
 
 ### Fixes
 - **todo-commithash** Don't show update available prompt if the found version is the same as the running app version